services/serve: move serve logic from starport/cmd
to the new services/serve package.
